Rework Configuration, Part 2
* Redo parts of SettingsUI:
  * Move it into a module
  * Remove all global usage
  * Disable mixmode changing when remote
* Move ICONFILE, MIXMODE, PAUSED into
  CONFIG removing three more globals
* Change all active/passive -> oldest/newst mixmode
* Provide new /index.txt via webserver which is
  content of text file (UTF-8 encoded!)
* On Windows, use ini format for settings, not registry
